+++ starky2/src/lib.rs
@@ -5,6 +5,7 @@
 
 pub mod config;
 pub mod constraint_consumer;
+pub mod cross_table_lookups;
 mod get_challenges;
 pub mod mock_stark;
 pub mod permutation;
